requirements.txt
Django==4.1.6
django-environ==0.9.0
django-mptt==0.14.0
psycopg2==2.9.5
